+++ vista/algorithms/background_removal/__init__.py
@@ -1,7 +1,16 @@
 """Background removal algorithms for VISTA"""
 from .godec import godec
 from .robust_pca import run_robust_pca
+from .static_median_background_removal import static_median_background_removal
+from .static_subspace_background_removal import static_subspace_background_removal
 from .subspace_background_removal import subspace_background_removal
 from .temporal_median import TemporalMedian
 
-__all__ = ['godec', 'run_robust_pca', 'subspace_background_removal', 'TemporalMedian']
+__all__ = [
+    'godec',
+    'run_robust_pca',
+    'static_median_background_removal',
+    'static_subspace_background_removal',
+    'subspace_background_removal',
+    'TemporalMedian',
+]